Moved AppEngineFacet from com.google.cloud.tools.eclipse.appengine.newproject to com.google.cloud.tools.eclipse.appengine.facets
