+++ plugins/com.google.cloud.tools.eclipse.appengine.facets/META-INF/MANIFEST.MF
@@ -11,8 +11,13 @@ Require-Bundle: com.google.cloud.tools.app.lib;bundle-version="0.1.0",
  com.google.guava;bundle-version="15.0.0",
  org.eclipse.core.resources,
  org.eclipse.jdt.core,
- org.eclipse.jst.j2ee
+ org.eclipse.jst.j2ee,
+ org.eclipse.jst.j2ee.web
 Import-Package: com.google.cloud.tools.eclipse.util,
  org.eclipse.core.runtime;version="3.5.0",
+ org.eclipse.jface.preference,
+ org.eclipse.jst.server.core,
  org.eclipse.wst.common.project.facet.core,
- org.eclipse.wst.common.project.facet.core.runtime
+ org.eclipse.wst.common.project.facet.core.runtime,
+ org.eclipse.wst.server.core
+Export-Package: com.google.cloud.tools.eclipse.appengine.facets
