+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject/META-INF/MANIFEST.MF
@@ -18,7 +18,8 @@ Require-Bundle: com.google.cloud.tools.app.lib;bundle-version="0.1.0",
  org.eclipse.jst.j2ee.web,
  com.google.cloud.tools.eclipse.sdk;bundle-version="0.1.0",
  com.google.cloud.tools.eclipse.sdk.ui;bundle-version="0.1.0"
-Import-Package: com.google.cloud.tools.eclipse.appengine.localserver.runtime,
+Import-Package: com.google.cloud.tools.eclipse.appengine.facets,
+ com.google.cloud.tools.eclipse.appengine.localserver.runtime,
  org.eclipse.core.runtime;version="3.5.0",
  org.eclipse.jdt.core,
  org.eclipse.jst.common.project.facet.core,
